requirements-lock.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
akracer==0.0.14
akshare==1.18.37
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
curl_cffi==0.14.0
cut-cross-entropy==25.1.1
datasets==4.3.0
decorator==5.2.1
diffusers==0.37.0
dill==0.4.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.3.2
hf_transfer==0.1.9
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonpath==0.82.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.16
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
peft==0.18.1
pillow==12.0.0
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
py-mini-racer==0.6.0
pyarrow==23.0.1
pycparser==3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
sentencepiece==0.2.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.13.1
tabulate==0.10.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.6.0+cu124
torchao==0.16.0
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
transformers==5.2.0
triton==3.2.0
trl==0.24.0
typeguard==4.5.1
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tyro==1.0.8
unsloth @ git+https://github.com/unslothai/unsloth.git@e70dd1d171f4d605dc2928ff499518560018d188
unsloth_zoo==2026.3.2
urllib3==2.6.3
webencodings==0.5.1
wheel==0.46.3
xformers @ https://download.pytorch.org/whl/cu124/xformers-0.0.29.post3-cp312-cp312-manylinux_2_28_x86_64.whl#sha256=08fa92f3e06372c4ce2a5306c54ae3d4a3a399fc7e24e02aac3761112ec3aeed
xlrd==2.0.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
